Add streaming to events list
